Move CallNodeInfo interface back into types/.
