+++ src/types/profile-derived.js
@@ -24,8 +24,65 @@ import type {
   Tid,
 } from './profile';
 import type { IndexedArray } from './utils';
-export type { CallNodeInfo } from '../profile-logic/profile-data';
+
 export type IndexIntoCallNodeTable = number;
+export type IndexIntoInvertedOrdering = number;
+export type CallNodeHandle = number;
+
+export interface CallNodeInfo {
+  isInverted(): boolean;
+  asInverted(): CallNodeInfoInverted | null;
+  getNonInvertedCallNodeTable(): CallNodeTable;
+  getStackIndexToNonInvertedCallNodeIndex(): Int32Array;
+
+  getCallNodePathFromIndex(
+    callNodeIndex: IndexIntoCallNodeTable | null
+  ): CallNodePath;
+
+  // Returns a list of CallNodeIndex from CallNodePaths.
+  getCallNodeIndicesFromPaths(
+    callNodePaths: CallNodePath[]
+  ): Array<IndexIntoCallNodeTable | null>;
+
+  // This function returns a CallNodeIndex from a CallNodePath.
+  getCallNodeIndexFromPath(
+    callNodePath: CallNodePath
+  ): IndexIntoCallNodeTable | null;
+
+  // Returns the CallNodeIndex that matches the function `func` and whose parent's
+  // CallNodeIndex is `parent`.
+  getCallNodeIndexFromParentAndFunc(
+    parent: IndexIntoCallNodeTable | -1,
+    func: IndexIntoFuncTable
+  ): IndexIntoCallNodeTable | null;
+
+  getParentCallNodeIndex(
+    callNodeIndex: IndexIntoCallNodeTable
+  ): IndexIntoCallNodeTable | null;
+
+  funcForNode(callNodeIndex: IndexIntoCallNodeTable): IndexIntoFuncTable;
+  categoryForNode(callNodeIndex: IndexIntoCallNodeTable): IndexIntoCategoryList;
+  subcategoryForNode(
+    callNodeIndex: IndexIntoCallNodeTable
+  ): IndexIntoCategoryList;
+  innerWindowIDForNode(
+    callNodeIndex: IndexIntoCallNodeTable
+  ): IndexIntoCategoryList;
+  depthForNode(callNodeIndex: IndexIntoCallNodeTable): number;
+  sourceFramesInlinedIntoSymbolForNode(
+    callNodeIndex: IndexIntoCallNodeTable
+  ): IndexIntoNativeSymbolTable | -1 | null;
+}
+
+export interface CallNodeInfoInverted extends CallNodeInfo {
+  getChildren(nodeHandle: CallNodeHandle): CallNodeHandle[];
+  getOrderedSelfNodes(): Uint32Array;
+  getOrderingIndexForSelfNode(): Uint32Array;
+  getRootCount(): number;
+  getOrderingIndexRangeForNode(
+    nodeHandle: CallNodeHandle
+  ): [IndexIntoInvertedOrdering, IndexIntoInvertedOrdering];
+}
 
 /**
  * Contains a table of function call information that represents the stacks of what
